+++ docs/ReleaseNotes.md
@@ -1,6 +1,7 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 0.1.9.25 Refactor
+# 0.1.9.26 Download PopuUp and Services Refactor
 
-- DatabaseService.savePictureFromUri : now has a onSave Callback.
-- <AppCamera />: Now adds picture amount feedback only if onSave callback is called.
+- REST Service created, to handle server calls and create errors specific to the calls
+- Download Service created, to handle logics behind data downloading.
+- Data Processing Service create, to focus on any type of data processing.